doc i0 minimal statement of project purpose
This readme file indicates that this project is for the rockforddev website.

I'm adding this file as git(hub?) requires a repository to have some file before we can clone / fork it.

I attest that I release this specific commit's content under maximally permissive terms, aka CC0 or public domain (where legally available). Future commits will be under the terms of the license that we agree to.